+++ src/julia.c
@@ -17,8 +17,8 @@
 int	iter_julia_1(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = dot.c_im;
 	dot.z_re = dot.c_re;
@@ -29,8 +29,8 @@ int	iter_julia_1(t_dot dot, t_env *env)
 			break ;
 		tmp_re = dot.z_re;
 		tmp_im = dot.z_im;
-		dot.z_re = tmp_re + ((float)env->mouse_x) / WIN_WIDTH;
-		dot.z_im = tmp_im + ((float)env->mouse_y) / WIN_HEIGHT;
+		dot.z_re = tmp_re + ((double)env->mouse_x) / WIN_WIDTH;
+		dot.z_im = tmp_im + ((double)env->mouse_y) / WIN_HEIGHT;
 	}
 	return (iter);
 }
@@ -38,8 +38,8 @@ int	iter_julia_1(t_dot dot, t_env *env)
 int	iter_julia_2(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = dot.c_im;
 	dot.z_re = dot.c_re;
@@ -50,8 +50,8 @@ int	iter_julia_2(t_dot dot, t_env *env)
 			break ;
 		tmp_re = POW2(dot.z_re) - POW2(dot.z_im);
 		tmp_im = 2 * dot.z_re * dot.z_im;
-		dot.z_re = tmp_re + ((float)env->mouse_x) / WIN_WIDTH;
-		dot.z_im = tmp_im + ((float)env->mouse_y) / WIN_HEIGHT;
+		dot.z_re = tmp_re + ((double)env->mouse_x) / WIN_WIDTH;
+		dot.z_im = tmp_im + ((double)env->mouse_y) / WIN_HEIGHT;
 	}
 	return (iter);
 }
@@ -59,8 +59,8 @@ int	iter_julia_2(t_dot dot, t_env *env)
 int	iter_julia_3(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = dot.c_im;
 	dot.z_re = dot.c_re;
@@ -71,8 +71,8 @@ int	iter_julia_3(t_dot dot, t_env *env)
 			break ;
 		tmp_re = POW3(dot.z_re) - 3 * dot.z_re * POW2(dot.z_im);
 		tmp_im = 3 * POW2(dot.z_re) * dot.z_im - POW3(dot.z_im);
-		dot.z_re = tmp_re + ((float)env->mouse_x) / WIN_WIDTH;
-		dot.z_im = tmp_im + ((float)env->mouse_y) / WIN_HEIGHT;
+		dot.z_re = tmp_re + ((double)env->mouse_x) / WIN_WIDTH;
+		dot.z_im = tmp_im + ((double)env->mouse_y) / WIN_HEIGHT;
 	}
 	return (iter);
 }
@@ -80,8 +80,8 @@ int	iter_julia_3(t_dot dot, t_env *env)
 int	iter_julia_4(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = dot.c_im;
 	dot.z_re = dot.c_re;
@@ -93,8 +93,8 @@ int	iter_julia_4(t_dot dot, t_env *env)
 		tmp_re = POW4(dot.z_re)
 		+ POW4(dot.z_im) - 6 * POW2(dot.z_re) * POW2(dot.z_im);
 		tmp_im = 4 * dot.z_im * POW3(dot.z_re) - 4 * POW3(dot.z_im) * dot.z_re;
-		dot.z_re = tmp_re + ((float)env->mouse_x) / WIN_WIDTH;
-		dot.z_im = tmp_im + ((float)env->mouse_y) / WIN_HEIGHT;
+		dot.z_re = tmp_re + ((double)env->mouse_x) / WIN_WIDTH;
+		dot.z_im = tmp_im + ((double)env->mouse_y) / WIN_HEIGHT;
 	}
 	return (iter);
 }
@@ -102,8 +102,8 @@ int	iter_julia_4(t_dot dot, t_env *env)
 int	iter_julia_5(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = dot.c_im;
 	dot.z_re = dot.c_re;
@@ -114,8 +114,8 @@ int	iter_julia_5(t_dot dot, t_env *env)
 			break ;
 		tmp_re = 5 * POW4(dot.z_im) * dot.z_re - 10 * POW2(dot.z_im) * POW3(dot.z_re) + POW5(dot.z_re);
 		tmp_im = POW5(dot.z_im) - 10 * POW3(dot.z_im) * POW2(dot.z_re) + 5 * POW4(dot.z_re) * dot.z_im;
-		dot.z_re = tmp_re + ((float)env->mouse_x) / WIN_WIDTH;
-		dot.z_im = tmp_im + ((float)env->mouse_y) / WIN_HEIGHT;
+		dot.z_re = tmp_re + ((double)env->mouse_x) / WIN_WIDTH;
+		dot.z_im = tmp_im + ((double)env->mouse_y) / WIN_HEIGHT;
 	}
 	return (iter);
 }
