+++ src/mandelbar.c
@@ -15,8 +15,8 @@
 int	iter_mandelbar_1(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = 0;
 	dot.z_re = 0;
@@ -36,8 +36,8 @@ int	iter_mandelbar_1(t_dot dot, t_env *env)
 int	iter_mandelbar_2(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = 0;
 	dot.z_re = 0;
@@ -57,8 +57,8 @@ int	iter_mandelbar_2(t_dot dot, t_env *env)
 int	iter_mandelbar_3(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = 0;
 	dot.z_re = 0;
@@ -78,8 +78,8 @@ int	iter_mandelbar_3(t_dot dot, t_env *env)
 int	iter_mandelbar_4(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = 0;
 	dot.z_re = 0;
@@ -100,8 +100,8 @@ int	iter_mandelbar_4(t_dot dot, t_env *env)
 int	iter_mandelbar_5(t_dot dot, t_env *env)
 {
 	int		iter;
-	float	tmp_re;
-	float	tmp_im;
+	double	tmp_re;
+	double	tmp_im;
 
 	dot.z_im = 0;
 	dot.z_re = 0;
